+++ Tests/Objects/IGListTestCollectionViewLayout.h
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+#import <UIKit/UIKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+/// Layout that 1) creates all attributes regardless of size, and 2) positions them with origin (0,0)
+@interface IGListTestCollectionViewLayout : UICollectionViewLayout
+
+@end
+
+NS_ASSUME_NONNULL_END